[Glitch] Convert `entrypoints/sign_up` to Typescript
Port 4527e012daa61dc07258eaedf3e03179b59fca8d to glitch-soc
